// src/types/generated/MatchFetchPlan.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MatchFetchDecision } from "./MatchFetchDecision";

export type MatchFetchPlan = { 
/**
 * Match ids to fetch, in selection (priority) order.
 */
to_fetch: Array<string>, decisions: Array<MatchFetchDecision>, batch_limit: number, selected_count: number, skipped_count: number, };